requirements.txt
# Water Leak Localization - Dependencies
# Python 3.10+

# Core ML
torch>=2.0.0

# Graph & Numerics
networkx>=3.0
numpy>=1.24.0
scipy>=1.10.0

# Visualization
matplotlib>=3.7.0
plotly>=5.15.0

# Web App
streamlit>=1.25.0

# Configuration
pyyaml>=6.0

# ML Utilities
scikit-learn>=1.3.0

# Progress bars
tqdm>=4.65.0
